Create Iceberg Limit Orders
This script is a plug in for placing "Iceberg Orders" learn more in the forum:

https://discuss.tradewave.net/t/iceberg-limit-orders-v1-0-beta-litepresence/142

You can see what's happening best on about 75 minutes of 1m data, then run another test on 24 hours of data. It always buys the bottom of the price channel and sells the top of the price channel. Traditional "Market Orders" do the OPPOSITE when traded live; they buy the top of the price channel and sell the bottom. 

Iceberg orders; depending on volatility might save you 0.5% per trade. 


Its trade decision as coded is random and would need your trading code to replace it. Its the order placement within the price channel that this plug in addresses.

It could be supplemented with an indicator like MFI for handling larger volume transactions.

I'm approximating spread based on previous 1m candle vwap and standard deviation. Then I'm placing a gaussian stack of orders in the most likely place to be bought (folded on the wide side). The same could be done on 5, 10 or 15m candles; or with longer vwap and std periods depending on risk preference. You could even pass different risk preferences depending upon the percieved urgency of the order type... ie bot says "buy criteria 1" and passes buy to iceberg with short vwap and std periods. then the bot sells with "criteria 3" and it might pass longer vwap and std periods because the criteria 3 sale is less urgent.


NOT LIVE TESTED USE AT OWN RISK 
MUST BE TUNED LIVE WITH SMALL FUNDS BEFORE DEPLOYING